consolidate API classes (so that we can seal the families)
