+++ api/src/main/scala/org/ensime/config.scala
