--- api/src/main/scala/org/ensime/server/ConnectionInfo.scala
@@ -1,11 +1,0 @@
-package org.ensime.server
-
-case class EnsimeImplementation(
-  name: String
-)
-case class ConnectionInfo(
-  pid: Option[Int] = None,
-  implementation: EnsimeImplementation = EnsimeImplementation("ENSIME"),
-  // Please also update changelog in SwankProtocol.scala
-  version: String = "0.8.14"
-)
