--- api/src/main/scala/org/ensime/util/FileRange.scala
@@ -1,3 +1,0 @@
-package org.ensime.util
-
-case class FileRange(file: String, start: Int, end: Int)
